requirements.txt
asgiref==3.7.2
blessings==1.7
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
cmd-utils==1.0.0
configparser==5.3.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography==41.0.1
cvxpy-base==1.3.2
cvxpy-leximin==0.4.4
cycler==0.11.0
DateTime==5.1
defusedxml==0.7.1
dicttools @ git+https://github.com/trzemecki/dicttools.git@39c19d9a5ecc965a58eed3eab18ac3e5e342fca2
Django==4.2.2
django-admin-relation-links==0.2.5
django-allauth==0.54.0
django-daterange-filter==1.3.0
django-filter==23.2
django-guardian==2.4.0
django-hijack==3.4.1
django-rangefilter==0.9.1
django-rq==2.8.1
django-safedelete==1.3.2
django-user-agents==0.4.0
execnet==1.9.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
guardian==0.2.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
meld3==2.0.1
mip==1.15.0
mock==5.0.2
more-itertools==9.1.0
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
path==16.6.0
path.py==12.5.0
Pillow==9.5.0
pluggy==1.2.0
prtpy==0.8.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygit2==1.12.1
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.4.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
repackage==0.7.3
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.15.1
rsa==4.9
SafeDelete==0.0.2
scipy==1.11.0
scs==3.2.3
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
termcolor==2.3.0
ua-parser==0.16.1
uritools==4.0.1
urllib3==1.26.16
user-agents==2.2.0
watchdog==3.0.0
Werkzeug==2.3.6
zope.interface==6.0
